pyproject.toml
[project]
name = "mle-bench-purple-agent"
version = "0.1.0"
description = "MLE-bench purple agent - solves Kaggle competitions via LangGraph + OpenRouter"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.3.0",
    "requests>=2.32.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
